done v1

+++ include/kernels.cuh
@@ -3,24 +3,24 @@
 #include <cstddef>
 
 template <typename T>
-void launch_gemm_kernel_v0(size_t m, size_t n, size_t k, T const* alpha,
-                            T const* A, size_t lda, T const* B, size_t ldb,
-                            T const* beta, T* C, size_t ldc,
-                            cudaStream_t stream);
+void launch_gemm_kernel_v0(int m, int n, int k, T const* alpha,
+                           T const* A, int lda, T const* B, int ldb,
+                           T const* beta, T* C, int ldc,
+                           cudaStream_t stream);
 
 template <typename T>
-void launch_gemm_cublas(size_t m, size_t n, size_t k, T const* alpha,
-                        T const* A, size_t lda, T const* B, size_t ldb,
-                        T const* beta, T* C, size_t ldc,
+void launch_gemm_cublas(int m, int n, int k, T const* alpha,
+                        T const* A, int lda, T const* B, int ldb,
+                        T const* beta, T* C, int ldc,
                         cudaStream_t stream);
 
 template <typename T>
-void launch_gemm_kernel_v1(size_t m, size_t n, size_t k, T const* alpha,
-                            T const* A, size_t lda, T const* B, size_t ldb,
-                            T const* beta, T* C, size_t ldc,
-                            cudaStream_t stream);
+void launch_gemm_kernel_v1(int m, int n, int k, T const* alpha,
+                           T const* A, int lda, T const* B, int ldb,
+                           T const* beta, T* C, int ldc,
+                           cudaStream_t stream);
 template <typename T>
-void launch_gemm_kernel_v2(size_t m, size_t n, size_t k, T const* alpha,
-                            T const* A, size_t lda, T const* B, size_t ldb,
-                            T const* beta, T* C, size_t ldc,
-                            cudaStream_t stream);
+void launch_gemm_kernel_v2(int m, int n, int k, T const* alpha,
+                           T const* A, int lda, T const* B, int ldb,
+                           T const* beta, T* C, int ldc,
+                           cudaStream_t stream);
